Updated to C#6
Suprisingly, this fixes an error reported by the installer, as the C#6 compiler adds a reference to an assembly which is only referenced in a typeof expression in an attribute. The C#5 compiler didn't add the reference (the typeof is serialised into the custom assembly, rather than being a typical type reference). The installer sees the type referenced in the attribute, and (with C#5) sees that an assembly reference is "missing" and reports an error. Since C#6 adds the reference, all is good.

Fixes #6